Making weather response to days converter

+++ data/src/main/java/com/helpfulapps/data/api/weather/model/Forecast.kt
@@ -6,13 +6,14 @@ data class Forecast(
     val clouds: Clouds?,
     val rain: Rain?,
     val snow: Snow?,
-    val dt: Int,
+    val dt: Long,
     val dt_txt: String,
     val main: Main,
     val sys: Sys,
     val weather: List<Weather>,
     val wind: Wind?
 ) {
+
     fun toDbModel() =
         WeatherDbModel(
             timestamp = dt,
